history extraction and graph creation done...?
I think that the code for pulling data from the history file and
putting it into a graph is finished, and my tests come out OK,
but more rigorous testing might be good. Currently, only Firefox
is supported, and I think Chromium is the only other one we can
add to that (IE is a binary blob, Safari uses plist (?)).