OK, NOW it's working. Call this good for now
So I messed up the Chromium query, due to a misunderstanding in how
the history is stored. I assumed that it would track a more exact
path from the same open tab, similar to the way I thought Firefox
does it (I think I need to check on that) rather than just putting
0 for the last page visited. Chromium and Firefox are now set so
they go based on chronological order rather than just asking what
the page before this one was.