GUI size/placement/behavior mostly complete
The GUI is almost done. The widgets are placed properly, both of
the buttons perform their proper actions, an error dialog pops up
when histparse throws an exception, the window cannot be resized,
and the entry widget starts with focus. A function was also added
to histparse.py to act as the main function to get the graph data
from a file path.